Refactor dashboard assets and refresh dark theme
