+++ public/index.php
@@ -224,2398 +224,6 @@ $streamInterval = getStatusStreamInterval();
     <p>Miłego dnia! 😊</p>
   </footer>
 
-  <script>
-    (function () {
-      const STATUS_JSON_ENDPOINT = '?status=json';
-      const STATUS_STREAM_ENDPOINT = '?status=stream';
-      const STATUS_HISTORY_ENDPOINT = '?status=history';
-
-      const readStreamInterval = () => {
-        if (!document.body || !document.body.dataset) {
-          return null;
-        }
-
-        const raw = document.body.dataset.streamInterval;
-        if (typeof raw !== 'string' || raw.trim() === '') {
-          return null;
-        }
-
-        const parsed = Number.parseInt(raw, 10);
-        if (!Number.isFinite(parsed) || parsed <= 0) {
-          return null;
-        }
-
-        return parsed;
-      };
-
-      const STREAM_INTERVAL_SECONDS = readStreamInterval();
-
-      const REFRESH_INTERVAL = (() => {
-        const fallback = 15000;
-
-        if (typeof STREAM_INTERVAL_SECONDS === 'number'
-          && Number.isFinite(STREAM_INTERVAL_SECONDS)
-          && STREAM_INTERVAL_SECONDS > 0
-        ) {
-          return Math.max(STREAM_INTERVAL_SECONDS * 1000, fallback);
-        }
-
-        return fallback;
-      })();
-      const STREAM_RECONNECT_DELAY = 5000;
-      const HISTORY_DEFAULT_LIMIT = 360;
-      const HISTORY_DEFAULT_METRIC = 'cpuTemperature';
-      const SHELLY_LIST_ENDPOINT = '?shelly=list';
-      const SHELLY_COMMAND_ENDPOINT = '?shelly=command';
-      const SHELLY_AUTO_REFRESH_INTERVAL = 1000;
-      const STATUS_TAB_ID = 'status';
-      const SHELLY_TAB_ID = 'shelly';
-      const TAB_STORAGE_KEY = 'dashboard-active-tab';
-      const HISTORY_METRIC_STORAGE_KEY = 'dashboard-history-metric';
-      const SHELLY_CONFIG_ERROR_MESSAGE = 'Nie udało się wczytać konfiguracji urządzeń Shelly. Sprawdź ustawienia w pliku config/shelly.php.';
-      const SHELLY_CONFIG_ERROR_HINT = 'Panel Shelly jest niedostępny do czasu poprawienia konfiguracji.';
-      const CSRF_TOKEN = document.body && document.body.dataset ? document.body.dataset.csrfToken || '' : '';
-
-      const elements = {
-        time: document.querySelector('[data-role="server-time"]'),
-        cpuTemperature: document.querySelector('[data-role="cpu-temperature"]'),
-        memoryUsage: document.querySelector('[data-role="memory-usage"]'),
-        diskUsage: document.querySelector('[data-role="disk-usage"]'),
-        systemLoad: document.querySelector('[data-role="system-load"]'),
-        uptime: document.querySelector('[data-role="uptime"]'),
-        serviceList: document.querySelector('[data-role="service-list"]'),
-        refreshLabel: document.querySelector('[data-role="refresh-label"]'),
-        refreshButton: document.querySelector('[data-role="refresh-button"]'),
-        refreshContainer: document.querySelector('[data-role="refresh-container"]'),
-        statusNote: document.querySelector('[data-role="status-note"]'),
-        historyContainer: document.querySelector('[data-role="history-container"]'),
-        historyChartWrapper: document.querySelector('[data-role="history-chart-wrapper"]'),
-        historyChart: document.querySelector('[data-role="history-chart"]'),
-        historyTitle: document.querySelector('[data-role="history-title"]'),
-        historyMetricButtons: document.querySelectorAll('[data-role="history-metric"]'),
-        historyEmpty: document.querySelector('[data-role="history-empty"]'),
-        themeToggle: document.querySelector('[data-role="theme-toggle"]'),
-        tabs: document.querySelector('[data-role="tabs"]'),
-        tabButtons: document.querySelectorAll('[data-role="tab"]'),
-        tabPanels: document.querySelectorAll('[data-role="tab-panel"]'),
-        shellyPanel: document.querySelector('[data-tab-panel="shelly"]'),
-        shellyList: document.querySelector('[data-role="shelly-list"]'),
-        shellyError: document.querySelector('[data-role="shelly-error"]'),
-        shellyMessage: document.querySelector('[data-role="shelly-message"]'),
-        shellyReload: document.querySelector('[data-role="shelly-reload"]'),
-        shellyLastUpdate: document.querySelector('[data-role="shelly-last-update"]'),
-      };
-
-      const shellyConfigError = Boolean(
-        elements.shellyPanel
-        && elements.shellyPanel.getAttribute('data-shelly-config-error') === 'true'
-      );
-
-      const THEME_STORAGE_KEY = 'theme-preference';
-
-      const readThemePreference = () => {
-        try {
-          if (typeof window.localStorage === 'undefined') {
-            return null;
-          }
-          return window.localStorage.getItem(THEME_STORAGE_KEY);
-        } catch (error) {
-          return null;
-        }
-      };
-
-      const writeThemePreference = (value) => {
-        try {
-          if (typeof window.localStorage === 'undefined') {
-            return;
-          }
-          window.localStorage.setItem(THEME_STORAGE_KEY, value);
-        } catch (error) {
-          // Ignorujemy błędy zapisu (np. tryb prywatny)
-        }
-      };
-
-      const updateThemeToggle = (isDark) => {
-        if (!elements.themeToggle) {
-          return;
-        }
-        elements.themeToggle.textContent = isDark ? 'Tryb ciemny włączony' : 'Włącz tryb ciemny';
-        elements.themeToggle.setAttribute('aria-pressed', isDark ? 'true' : 'false');
-      };
-
-      const applyThemePreference = (isDark, persist = false) => {
-        document.body.classList.toggle('theme-dark', isDark);
-        updateThemeToggle(isDark);
-        if (persist) {
-          writeThemePreference(isDark ? 'dark' : 'light');
-        }
-      };
-
-      const initializeTheme = () => {
-        const storedPreference = readThemePreference();
-        let isDark = storedPreference === 'dark';
-
-        if (storedPreference !== 'dark' && storedPreference !== 'light') {
-          if (window.matchMedia && typeof window.matchMedia === 'function') {
-            const prefersDark = window.matchMedia('(prefers-color-scheme: dark)');
-            if (prefersDark && typeof prefersDark.matches === 'boolean') {
-              isDark = prefersDark.matches;
-            }
-          } else {
-            isDark = document.body.classList.contains('theme-dark');
-          }
-        }
-
-        applyThemePreference(isDark, storedPreference !== 'dark' && storedPreference !== 'light');
-      };
-
-      let currentTab = STATUS_TAB_ID;
-
-      const readActiveTabPreference = () => {
-        try {
-          if (typeof window.localStorage === 'undefined') {
-            return null;
-          }
-          const stored = window.localStorage.getItem(TAB_STORAGE_KEY);
-          return typeof stored === 'string' && stored.trim() !== '' ? stored : null;
-        } catch (error) {
-          return null;
-        }
-      };
-
-      const writeActiveTabPreference = (value) => {
-        try {
-          if (typeof window.localStorage === 'undefined') {
-            return;
-          }
-          window.localStorage.setItem(TAB_STORAGE_KEY, value);
-        } catch (error) {
-          // Ignorujemy błędy zapisu.
-        }
-      };
-
-      const readHistoryMetricPreference = () => {
-        try {
-          if (typeof window.localStorage === 'undefined') {
-            return null;
-          }
-          const stored = window.localStorage.getItem(HISTORY_METRIC_STORAGE_KEY);
-          return typeof stored === 'string' && stored.trim() !== '' ? stored : null;
-        } catch (error) {
-          return null;
-        }
-      };
-
-      const writeHistoryMetricPreference = (value) => {
-        try {
-          if (typeof window.localStorage === 'undefined') {
-            return;
-          }
-          window.localStorage.setItem(HISTORY_METRIC_STORAGE_KEY, value);
-        } catch (error) {
-          // Ignorujemy błędy zapisu (np. tryb prywatny).
-        }
-      };
-
-      const setActiveTab = (tabId, persist = false) => {
-        const tabButtons = Array.from(elements.tabButtons || []);
-        const tabPanels = Array.from(elements.tabPanels || []);
-
-        if (tabButtons.length === 0 || tabPanels.length === 0) {
-          currentTab = STATUS_TAB_ID;
-          return currentTab;
-        }
-
-        const normalized = typeof tabId === 'string' && tabId.trim() !== '' ? tabId.trim() : STATUS_TAB_ID;
-
-        let found = false;
-
-        tabButtons.forEach((button) => {
-          const target = button.getAttribute('data-tab');
-          const isActive = target === normalized;
-          if (isActive) {
-            found = true;
-          }
-          button.classList.toggle('is-active', isActive);
-          button.setAttribute('aria-selected', isActive ? 'true' : 'false');
-          button.setAttribute('tabindex', isActive ? '0' : '-1');
-        });
-
-        tabPanels.forEach((panel) => {
-          const target = panel.getAttribute('data-tab-panel');
-          const isActive = target === normalized;
-          panel.classList.toggle('is-active', isActive);
-          panel.hidden = !isActive;
-        });
-
-        if (!found && normalized !== STATUS_TAB_ID) {
-          return setActiveTab(STATUS_TAB_ID, persist);
-        }
-
-        const finalTab = found ? normalized : STATUS_TAB_ID;
-        currentTab = finalTab;
-
-        if (persist) {
-          writeActiveTabPreference(finalTab);
-        }
-
-        return finalTab;
-      };
-
-      const initializeTabs = () => {
-        const tabButtons = Array.from(elements.tabButtons || []);
-        if (tabButtons.length === 0) {
-          return;
-        }
-
-        tabButtons.forEach((button) => {
-          button.addEventListener('click', () => {
-            const target = button.getAttribute('data-tab');
-            if (!target) {
-              return;
-            }
-
-            const active = setActiveTab(target, true);
-            handleTabActivation(active, true);
-          });
-        });
-
-        const stored = readActiveTabPreference();
-        const active = setActiveTab(stored || STATUS_TAB_ID, false);
-        handleTabActivation(active, false);
-      };
-
-      const supportsFetch = typeof window.fetch === 'function';
-      const supportsEventSource = typeof window.EventSource === 'function';
-
-      const shellyState = {
-        devices: [],
-        loading: false,
-        error: null,
-        hasErrors: false,
-        pending: new Set(),
-        lastUpdated: null,
-        lastAttempt: null,
-        initialized: false,
-        requestId: 0,
-      };
-
-      if (shellyConfigError) {
-        shellyState.error = SHELLY_CONFIG_ERROR_MESSAGE;
-        shellyState.initialized = true;
-      }
-
-      let shellyRefreshTimer = null;
-
-      const markShellyListIdle = () => {
-        const container = elements.shellyList;
-        if (!container) {
-          return;
-        }
-
-        container.classList.remove('is-refreshing');
-        if (!shellyState.loading) {
-          container.classList.remove('is-loading');
-        }
-        container.setAttribute('aria-busy', 'false');
-      };
-
-      const haveShellyDevicesChanged = (previousDevices, nextDevices) => {
-        if (!Array.isArray(previousDevices) || !Array.isArray(nextDevices)) {
-          return true;
-        }
-
-        if (previousDevices.length !== nextDevices.length) {
-          return true;
-        }
-
-        const previousById = new Map();
-        previousDevices.forEach((device) => {
-          if (device && typeof device.id === 'string') {
-            previousById.set(device.id, device);
-          }
-        });
-
-        if (previousById.size !== nextDevices.length) {
-          return true;
-        }
-
-        for (const device of nextDevices) {
-          if (!device || typeof device.id !== 'string') {
-            return true;
-          }
-
-          const previous = previousById.get(device.id);
-          if (!previous) {
-            return true;
-          }
-
-          if (
-            previous.label !== device.label
-            || previous.state !== device.state
-            || previous.description !== device.description
-            || previous.ok !== device.ok
-            || previous.error !== device.error
-          ) {
-            return true;
-          }
-        }
-
-        return false;
-      };
-
-      const setShellyError = (message) => {
-        if (!elements.shellyError) {
-          return;
-        }
-
-        if (typeof message === 'string' && message.trim() !== '') {
-          elements.shellyError.textContent = message;
-          elements.shellyError.hidden = false;
-        } else {
-          elements.shellyError.textContent = '';
-          elements.shellyError.hidden = true;
-        }
-      };
-
-      const setShellyMessage = (message) => {
-        if (!elements.shellyMessage) {
-          return;
-        }
-
-        if (typeof message === 'string' && message.trim() !== '') {
-          elements.shellyMessage.textContent = message;
-          elements.shellyMessage.hidden = false;
-        } else {
-          elements.shellyMessage.textContent = '';
-          elements.shellyMessage.hidden = true;
-        }
-      };
-
-      const updateShellyLastUpdate = () => {
-        if (!elements.shellyLastUpdate) {
-          return;
-        }
-
-        if (shellyState.loading) {
-          elements.shellyLastUpdate.textContent = 'Trwa ładowanie...';
-          return;
-        }
-
-        if (typeof shellyState.lastUpdated === 'string' && shellyState.lastUpdated.trim() !== '') {
-          const parsed = new Date(shellyState.lastUpdated);
-          const formatted = Number.isNaN(parsed.getTime())
-            ? shellyState.lastUpdated
-            : parsed.toLocaleString();
-          elements.shellyLastUpdate.textContent = `Ostatnia aktualizacja: ${formatted}`;
-          return;
-        }
-
-        if (
-          shellyState.error
-          && typeof shellyState.lastAttempt === 'string'
-          && shellyState.lastAttempt.trim() !== ''
-        ) {
-          const attempt = new Date(shellyState.lastAttempt);
-          const formattedAttempt = Number.isNaN(attempt.getTime())
-            ? shellyState.lastAttempt
-            : attempt.toLocaleString();
-          elements.shellyLastUpdate.textContent = `Ostatnia próba: ${formattedAttempt}`;
-          return;
-        }
-
-        elements.shellyLastUpdate.textContent = '';
-      };
-
-      const formatShellyStateLabel = (state) => {
-        switch (state) {
-          case 'on':
-            return 'Włączone';
-          case 'off':
-            return 'Wyłączone';
-          default:
-            return 'Nieznany stan';
-        }
-      };
-
-      const renderShellyDevices = () => {
-        const container = elements.shellyList;
-        if (!container) {
-          return;
-        }
-
-        const hasDevices = shellyState.devices.length > 0;
-
-        if (shellyState.loading && hasDevices) {
-          container.classList.remove('is-loading');
-          container.classList.add('is-refreshing');
-          container.setAttribute('aria-busy', 'true');
-          updateShellyLastUpdate();
-          return;
-        }
-
-        container.classList.remove('is-refreshing');
-        container.innerHTML = '';
-
-        if (shellyState.loading) {
-          container.classList.add('is-loading');
-        } else {
-          container.classList.remove('is-loading');
-        }
-
-        container.setAttribute('aria-busy', shellyState.loading ? 'true' : 'false');
-
-        if (!shellyState.initialized && !shellyState.loading && !shellyState.error) {
-          setShellyMessage('Przełącz na kartę „Shelly”, aby pobrać stan urządzeń.');
-        }
-
-        if (shellyState.loading) {
-          setShellyMessage('Ładowanie urządzeń Shelly...');
-          setShellyError(null);
-
-          const loader = document.createElement('div');
-          loader.className = 'shelly-loading';
-          loader.setAttribute('role', 'status');
-          loader.setAttribute('aria-live', 'polite');
-          loader.textContent = 'Ładowanie urządzeń Shelly...';
-          container.appendChild(loader);
-
-          updateShellyLastUpdate();
-          return;
-        }
-
-        if (shellyState.error) {
-          setShellyError(shellyState.error);
-        } else {
-          setShellyError(null);
-        }
-
-        if (shellyState.devices.length === 0) {
-          if (!shellyState.error && shellyState.initialized) {
-            setShellyMessage('Brak skonfigurowanych urządzeń Shelly.');
-
-            const empty = document.createElement('p');
-            empty.className = 'shelly-empty';
-            empty.textContent = 'Brak danych do wyświetlenia.';
-            container.appendChild(empty);
-          } else if (shellyState.error) {
-            setShellyMessage('');
-          }
-
-          updateShellyLastUpdate();
-          return;
-        }
-
-        if (!shellyState.error) {
-          if (shellyState.hasErrors) {
-            setShellyMessage('Część urządzeń zgłosiła błąd podczas odczytu stanu.');
-          } else {
-            setShellyMessage('');
-          }
-        }
-
-        shellyState.devices.forEach((device) => {
-          const card = document.createElement('article');
-          card.className = 'shelly-device';
-
-          if (!device.ok) {
-            card.classList.add('has-error');
-          }
-
-          if (shellyState.pending.has(device.id)) {
-            card.classList.add('is-busy');
-          }
-
-          const body = document.createElement('div');
-          body.className = 'shelly-device__body';
-
-          const title = document.createElement('h3');
-          title.className = 'shelly-device__title';
-          title.textContent = device.label || device.id;
-          body.appendChild(title);
-
-          const stateLine = document.createElement('p');
-          stateLine.className = 'shelly-device__state';
-          const stateValue = document.createElement('span');
-          stateValue.className = 'shelly-device__state-value';
-          stateValue.dataset.state = device.state;
-          stateValue.textContent = formatShellyStateLabel(device.state);
-          stateLine.textContent = 'Stan: ';
-          stateLine.appendChild(stateValue);
-          body.appendChild(stateLine);
-
-          if (device.description) {
-            const description = document.createElement('p');
-            description.className = 'shelly-device__description';
-            description.textContent = device.description;
-            body.appendChild(description);
-          }
-
-          card.appendChild(body);
-
-          const actions = document.createElement('div');
-          actions.className = 'shelly-device__actions';
-
-          const actionButton = document.createElement('button');
-          const nextAction = device.state === 'on' ? 'off' : 'on';
-          actionButton.type = 'button';
-          actionButton.dataset.role = 'shelly-action';
-          actionButton.dataset.device = device.id;
-          actionButton.dataset.action = nextAction;
-          actionButton.classList.add('shelly-device__action');
-          actionButton.setAttribute('aria-label', 'Przełącz zasilanie');
-          actionButton.setAttribute('aria-pressed', device.state === 'on' ? 'true' : 'false');
-          actionButton.title = nextAction === 'on' ? 'Włącz urządzenie' : 'Wyłącz urządzenie';
-
-          const icon = document.createElement('span');
-          icon.className = 'icon-power';
-          icon.setAttribute('aria-hidden', 'true');
-          const srText = document.createElement('span');
-          srText.className = 'sr-only';
-          srText.textContent = `${nextAction === 'on' ? 'Włącz' : 'Wyłącz'} urządzenie. Stan: ${formatShellyStateLabel(device.state)}.`;
-
-          actionButton.appendChild(icon);
-          actionButton.appendChild(srText);
-
-          if (shellyState.pending.has(device.id)) {
-            actionButton.disabled = true;
-          }
-
-          actions.appendChild(actionButton);
-
-          card.appendChild(actions);
-
-          if (device.error) {
-            const errorLine = document.createElement('p');
-            errorLine.className = 'shelly-device__error';
-            errorLine.textContent = `Błąd: ${device.error}`;
-            card.appendChild(errorLine);
-          }
-
-          container.appendChild(card);
-        });
-
-        updateShellyLastUpdate();
-      };
-
-      const loadShellyDevices = (force = false) => {
-        if (shellyConfigError) {
-          setShellyError(SHELLY_CONFIG_ERROR_MESSAGE);
-          setShellyMessage(SHELLY_CONFIG_ERROR_HINT);
-          return;
-        }
-
-        if (!supportsFetch) {
-          setShellyError('Sterowanie Shelly wymaga przeglądarki obsługującej funkcję fetch.');
-          return;
-        }
-
-        if (shellyState.loading && !force) {
-          return;
-        }
-
-        const errorBeforeRequest = shellyState.error;
-        const requestId = shellyState.requestId + 1;
-        shellyState.requestId = requestId;
-        shellyState.loading = true;
-        shellyState.error = null;
-        shellyState.initialized = true;
-        shellyState.lastAttempt = new Date().toISOString();
-        renderShellyDevices();
-
-        const shellyListHeaders = {
-          Accept: 'application/json',
-        };
-
-        if (CSRF_TOKEN) {
-          shellyListHeaders['X-CSRF-Token'] = CSRF_TOKEN;
-        }
-
-        fetch(SHELLY_LIST_ENDPOINT, {
-          method: 'GET',
-          headers: shellyListHeaders,
-          cache: 'no-store',
-          credentials: 'same-origin',
-        })
-          .then((response) => {
-            if (!response.ok) {
-              return response
-                .json()
-                .catch(() => null)
-                .then((data) => {
-                  let message = '';
-                  if (data && typeof data === 'object') {
-                    if (typeof data.error === 'string' && data.error.trim() !== '') {
-                      message = data.error.trim();
-                    }
-                    if (typeof data.message === 'string' && data.message.trim() !== '') {
-                      message = message !== '' ? `${message} - ${data.message.trim()}` : data.message.trim();
-                    }
-                    if (typeof data.description === 'string' && data.description.trim() !== '') {
-                      message = message !== '' ? `${message} - ${data.description.trim()}` : data.description.trim();
-                    }
-                  }
-
-                  if (message === '') {
-                    message = `HTTP ${response.status}`;
-                  }
-
-                  throw new Error(message);
-                });
-            }
-
-            return response.json();
-          })
-          .then((payload) => {
-            if (shellyState.requestId !== requestId) {
-              return;
-            }
-
-            const previousDevices = shellyState.devices;
-            const previousHasErrors = shellyState.hasErrors;
-            const previousError = errorBeforeRequest;
-            const devicesData = payload && typeof payload === 'object' && Array.isArray(payload.devices)
-              ? payload.devices
-              : [];
-
-            const normalizedDevices = devicesData.map((entry, index) => {
-              let id = '';
-              if (entry && typeof entry === 'object') {
-                if (typeof entry.id === 'string' && entry.id.trim() !== '') {
-                  id = entry.id.trim();
-                } else if (typeof entry.id === 'number' || typeof entry.id === 'boolean') {
-                  id = String(entry.id);
-                }
-              }
-
-              if (id === '') {
-                id = `device-${index + 1}`;
-              }
-
-              let label = id;
-              if (entry && typeof entry === 'object' && typeof entry.label === 'string' && entry.label.trim() !== '') {
-                label = entry.label.trim();
-              }
-
-              let state = 'unknown';
-              if (entry && typeof entry === 'object' && typeof entry.state === 'string') {
-                const normalizedState = entry.state.trim().toLowerCase();
-                if (normalizedState !== '') {
-                  state = normalizedState;
-                }
-              }
-
-              let description = '';
-              if (entry && typeof entry === 'object' && typeof entry.description === 'string') {
-                description = entry.description.trim();
-              }
-
-              let error = null;
-              if (entry && typeof entry === 'object' && typeof entry.error === 'string' && entry.error.trim() !== '') {
-                error = entry.error.trim();
-              }
-
-              let ok = true;
-              if (entry && typeof entry === 'object' && typeof entry.ok === 'boolean') {
-                ok = entry.ok;
-              } else if (error !== null) {
-                ok = false;
-              }
-
-              return {
-                id,
-                label,
-                state,
-                description,
-                ok: Boolean(ok),
-                error,
-              };
-            });
-
-            const devicesChanged = haveShellyDevicesChanged(previousDevices, normalizedDevices);
-            shellyState.devices = normalizedDevices;
-
-            let hasErrors = false;
-            if (payload && typeof payload === 'object' && typeof payload.hasErrors === 'boolean') {
-              hasErrors = payload.hasErrors;
-            }
-
-            shellyState.hasErrors = Boolean(hasErrors) || shellyState.devices.some((device) => !device.ok);
-
-            let generatedAt = null;
-            if (payload && typeof payload === 'object' && typeof payload.generatedAt === 'string') {
-              generatedAt = payload.generatedAt.trim();
-            }
-
-            shellyState.lastUpdated = generatedAt !== '' && generatedAt !== null ? generatedAt : shellyState.lastAttempt;
-            shellyState.error = null;
-            shellyState.loading = false;
-            if (devicesChanged || previousHasErrors !== shellyState.hasErrors || previousError !== shellyState.error) {
-              renderShellyDevices();
-            } else {
-              markShellyListIdle();
-            }
-            updateShellyLastUpdate();
-          })
-          .catch((error) => {
-            if (shellyState.requestId !== requestId) {
-              return;
-            }
-
-            shellyState.loading = false;
-            const hadDevices = Array.isArray(shellyState.devices) && shellyState.devices.length > 0;
-            shellyState.hasErrors = false;
-            const message = error && typeof error.message === 'string' && error.message.trim() !== ''
-              ? error.message.trim()
-              : 'Nie udało się pobrać listy urządzeń Shelly.';
-            shellyState.error = message;
-            if (hadDevices) {
-              setShellyError(message);
-              markShellyListIdle();
-            } else {
-              renderShellyDevices();
-            }
-            updateShellyLastUpdate();
-          });
-      };
-
-      const ensureShellyDataLoaded = (force = false) => {
-        if (shellyConfigError) {
-          return;
-        }
-
-        if (!supportsFetch) {
-          return;
-        }
-
-        loadShellyDevices(Boolean(force));
-      };
-
-      const stopShellyAutoRefresh = () => {
-        if (shellyRefreshTimer !== null) {
-          window.clearInterval(shellyRefreshTimer);
-          shellyRefreshTimer = null;
-        }
-      };
-
-      const startShellyAutoRefresh = () => {
-        if (shellyRefreshTimer !== null) {
-          return;
-        }
-
-        if (shellyConfigError || !supportsFetch) {
-          return;
-        }
-
-        if (typeof SHELLY_AUTO_REFRESH_INTERVAL !== 'number'
-          || !Number.isFinite(SHELLY_AUTO_REFRESH_INTERVAL)
-          || SHELLY_AUTO_REFRESH_INTERVAL <= 0
-        ) {
-          return;
-        }
-
-        if (typeof document.visibilityState === 'string'
-          && document.visibilityState !== 'visible'
-        ) {
-          return;
-        }
-
-        shellyRefreshTimer = window.setInterval(() => {
-          ensureShellyDataLoaded(false);
-        }, SHELLY_AUTO_REFRESH_INTERVAL);
-      };
-
-      const toggleShellyDevice = (deviceId, action) => {
-        if (shellyConfigError) {
-          setShellyError(SHELLY_CONFIG_ERROR_MESSAGE);
-          setShellyMessage(SHELLY_CONFIG_ERROR_HINT);
-          return;
-        }
-
-        if (!supportsFetch) {
-          setShellyError('Sterowanie Shelly wymaga przeglądarki obsługującej funkcję fetch.');
-          return;
-        }
-
-        const normalizedId = typeof deviceId === 'string' ? deviceId.trim() : '';
-        const normalizedAction = typeof action === 'string' ? action.trim().toLowerCase() : '';
-
-        if (normalizedId === '') {
-          return;
-        }
-
-        const device = shellyState.devices.find((item) => item.id === normalizedId);
-        let finalAction = normalizedAction === 'on' || normalizedAction === 'off'
-          ? normalizedAction
-          : null;
-
-        if (!finalAction && device) {
-          finalAction = device.state === 'on' ? 'off' : 'on';
-        }
-
-        if (finalAction !== 'on' && finalAction !== 'off') {
-          return;
-        }
-
-        if (shellyState.pending.has(normalizedId)) {
-          return;
-        }
-
-        shellyState.pending.add(normalizedId);
-        shellyState.error = null;
-        setShellyError(null);
-        renderShellyDevices();
-
-        const shellyCommandHeaders = {
-          'Content-Type': 'application/json',
-          Accept: 'application/json',
-        };
-
-        const shellyCommandPayload = {
-          device: normalizedId,
-          action: finalAction,
-        };
-
-        if (CSRF_TOKEN) {
-          shellyCommandHeaders['X-CSRF-Token'] = CSRF_TOKEN;
-          shellyCommandPayload.csrfToken = CSRF_TOKEN;
-        }
-
-        fetch(SHELLY_COMMAND_ENDPOINT, {
-          method: 'POST',
-          headers: shellyCommandHeaders,
-          body: JSON.stringify(shellyCommandPayload),
-          credentials: 'same-origin',
-        })
-          .then((response) => {
-            if (!response.ok) {
-              return response
-                .json()
-                .catch(() => null)
-                .then((data) => {
-                  const messageParts = [];
-                  if (data && typeof data === 'object') {
-                    if (typeof data.error === 'string' && data.error.trim() !== '') {
-                      messageParts.push(data.error.trim());
-                    }
-                    if (typeof data.description === 'string' && data.description.trim() !== '') {
-                      messageParts.push(data.description.trim());
-                    }
-                    if (typeof data.message === 'string' && data.message.trim() !== '') {
-                      messageParts.push(data.message.trim());
-                    }
-                  }
-
-                  if (messageParts.length === 0) {
-                    messageParts.push(`HTTP ${response.status}`);
-                  }
-
-                  throw new Error(messageParts.join(' - '));
-                });
-            }
-
-            return response.json();
-          })
-          .then((payload) => {
-            shellyState.pending.delete(normalizedId);
-
-            if (payload && typeof payload === 'object' && typeof payload.description === 'string') {
-              const trimmedDescription = payload.description.trim();
-              if (trimmedDescription !== '') {
-                setShellyMessage(trimmedDescription);
-              }
-            }
-
-            loadShellyDevices(true);
-          })
-          .catch((error) => {
-            shellyState.pending.delete(normalizedId);
-            const message = error && typeof error.message === 'string' && error.message.trim() !== ''
-              ? error.message.trim()
-              : 'Nie udało się wykonać polecenia Shelly.';
-            shellyState.error = message;
-            setShellyError(message);
-            renderShellyDevices();
-          });
-      };
-
-      const handleShellyActionClick = (event) => {
-        const target = event.target;
-        if (!target || typeof target.closest !== 'function') {
-          return;
-        }
-
-        const button = target.closest('[data-role="shelly-action"]');
-        if (!button) {
-          return;
-        }
-
-        const deviceId = button.getAttribute('data-device');
-        let action = button.getAttribute('data-action') || '';
-
-        if (!deviceId) {
-          return;
-        }
-
-        const device = shellyState.devices.find((item) => item.id === deviceId);
-        if (device) {
-          action = device.state === 'on' ? 'off' : 'on';
-        }
-
-        if (action !== 'on' && action !== 'off') {
-          return;
-        }
-
-        button.dataset.action = action;
-
-        event.preventDefault();
-        toggleShellyDevice(deviceId, action);
-      };
-
-      function handleTabActivation(tabId, fromUser) {
-        if (tabId === SHELLY_TAB_ID) {
-          if (shellyConfigError) {
-            stopShellyAutoRefresh();
-            setShellyError(SHELLY_CONFIG_ERROR_MESSAGE);
-            setShellyMessage(SHELLY_CONFIG_ERROR_HINT);
-            if (elements.shellyReload) {
-              elements.shellyReload.disabled = true;
-              elements.shellyReload.setAttribute('aria-disabled', 'true');
-            }
-            return;
-          }
-
-          if (!supportsFetch) {
-            stopShellyAutoRefresh();
-            setShellyError('Sterowanie Shelly wymaga przeglądarki obsługującej funkcję fetch.');
-            setShellyMessage('Sterowanie Shelly wymaga nowszej przeglądarki.');
-            if (elements.shellyReload) {
-              elements.shellyReload.disabled = true;
-              elements.shellyReload.setAttribute('aria-disabled', 'true');
-            }
-            return;
-          }
-
-          if (elements.shellyReload) {
-            elements.shellyReload.disabled = false;
-            elements.shellyReload.removeAttribute('aria-disabled');
-          }
-
-          ensureShellyDataLoaded(false);
-          startShellyAutoRefresh();
-        }
-        if (tabId !== SHELLY_TAB_ID) {
-          stopShellyAutoRefresh();
-        }
-      }
-
-      const historyState = {
-        enabled: null,
-        entries: [],
-        maxEntries: null,
-        limit: null,
-        chartSignature: null,
-        metric: HISTORY_DEFAULT_METRIC,
-
-      };
-
-      let refreshTimer = null;
-      let eventSource = null;
-      let reconnectTimer = null;
-      let isFetching = false;
-
-      const fallback = (value, fallbackValue = 'Brak danych') => {
-        if (typeof value === 'string') {
-          const trimmed = value.trim();
-          return trimmed !== '' ? trimmed : fallbackValue;
-        }
-        return fallbackValue;
-      };
-
-      const setFetchingState = (loading, manual = false, customText = null) => {
-        if (!elements.refreshContainer) {
-          return;
-        }
-
-        if (loading) {
-          elements.refreshContainer.classList.add('is-loading');
-        } else {
-          elements.refreshContainer.classList.remove('is-loading');
-        }
-
-        if (elements.refreshButton) {
-          elements.refreshButton.disabled = loading;
-        }
-
-        if (elements.refreshLabel) {
-          if (typeof customText === 'string' && customText.trim() !== '') {
-            elements.refreshLabel.textContent = customText;
-            return;
-          }
-
-          if (loading) {
-            const timestamp = new Date().toLocaleTimeString();
-            const prefix = manual ? 'Trwa ręczne odświeżanie' : 'Trwa odświeżanie';
-            elements.refreshLabel.textContent = `${prefix} (${timestamp})...`;
-          }
-        }
-      };
-
-      const setHistoryMessage = (message, hidden = false) => {
-        if (!elements.historyEmpty) {
-          return;
-        }
-
-        if (hidden) {
-          elements.historyEmpty.textContent = '';
-          elements.historyEmpty.hidden = true;
-          return;
-        }
-
-        elements.historyEmpty.textContent = message;
-        elements.historyEmpty.hidden = false;
-      };
-
-      const destroyHistoryChart = () => {
-        if (elements.historyChart) {
-          elements.historyChart.innerHTML = '';
-        }
-        historyState.chartSignature = null;
-      };
-
-      const getHistoryCapacity = () => {
-        if (typeof historyState.maxEntries === 'number' && historyState.maxEntries > 0) {
-          return historyState.maxEntries;
-        }
-        if (typeof historyState.limit === 'number' && historyState.limit > 0) {
-          return historyState.limit;
-        }
-        return HISTORY_DEFAULT_LIMIT;
-      };
-
-      const trimHistoryEntries = () => {
-        const capacity = getHistoryCapacity();
-        if (capacity > 0 && historyState.entries.length > capacity) {
-          historyState.entries = historyState.entries.slice(-capacity);
-        }
-      };
-
-      const CHART_VIEWBOX_WIDTH = 600;
-      const CHART_VIEWBOX_HEIGHT = 260;
-      const CHART_PADDING = { top: 24, right: 24, bottom: 36, left: 56 };
-      const CHART_GRID_LINES = 4;
-      const SVG_NS = 'http://www.w3.org/2000/svg';
-
-      const createSvgElement = (name, attributes = {}) => {
-        const element = document.createElementNS(SVG_NS, name);
-        Object.entries(attributes).forEach(([attribute, rawValue]) => {
-          if (rawValue === undefined || rawValue === null) {
-            return;
-          }
-          element.setAttribute(attribute, String(rawValue));
-        });
-        return element;
-      };
-
-      const formatTemperatureValue = (value) => {
-        if (typeof value !== 'number' || !Number.isFinite(value)) {
-          return '';
-        }
-        return `${value.toFixed(1)} °C`;
-      };
-
-      const formatPercentageValue = (value) => {
-        if (typeof value !== 'number' || !Number.isFinite(value)) {
-          return '';
-        }
-        return `${value.toFixed(1)} %`;
-      };
-
-      const formatLoadAverageValue = (value) => {
-        if (typeof value !== 'number' || !Number.isFinite(value)) {
-          return '';
-        }
-        return value.toFixed(2);
-      };
-
-      const formatHistoryEntryTime = (entry) => {
-        if (!entry || typeof entry !== 'object') {
-          return '—';
-        }
-
-        if (typeof entry.time === 'string' && entry.time.trim() !== '') {
-          return entry.time.trim();
-        }
-
-        if (typeof entry.generatedAt === 'string') {
-          const parsed = new Date(entry.generatedAt);
-          if (!Number.isNaN(parsed.getTime())) {
-            return parsed.toLocaleTimeString();
-          }
-        }
-
-        return '—';
-      };
-
-      const HISTORY_METRICS = [
-        {
-          id: 'cpuTemperature',
-          label: 'Temperatura CPU',
-          axisLabel: 'Temperatura CPU (°C)',
-          description: 'Wizualizacja historii temperatury CPU.',
-          formatValue: formatTemperatureValue,
-          getValue: (entry) => {
-            if (!entry || typeof entry !== 'object') {
-              return null;
-            }
-            const value = entry.cpuTemperatureValue;
-            return typeof value === 'number' && Number.isFinite(value) ? value : null;
-          },
-          getTooltipLabel: (entry, value) => {
-            if (entry && typeof entry.cpuTemperatureLabel === 'string' && entry.cpuTemperatureLabel.trim() !== '') {
-              return entry.cpuTemperatureLabel.trim();
-            }
-            const formatted = formatTemperatureValue(value);
-            if (formatted && formatted.trim() !== '') {
-              return formatted;
-            }
-            if (typeof value === 'number' && Number.isFinite(value)) {
-              return `${value.toFixed(1)} °C`;
-            }
-            return 'Brak danych';
-          },
-        },
-        {
-          id: 'memoryUsage',
-          label: 'Użycie pamięci RAM',
-          axisLabel: 'Użycie pamięci (%)',
-          description: 'Wizualizacja historii zużycia pamięci RAM.',
-          formatValue: formatPercentageValue,
-          getValue: (entry) => {
-            if (!entry || typeof entry !== 'object' || !entry.memoryUsage || typeof entry.memoryUsage !== 'object') {
-              return null;
-            }
-            const value = entry.memoryUsage.percentage;
-            return typeof value === 'number' && Number.isFinite(value) ? value : null;
-          },
-          getTooltipLabel: (entry, value) => {
-            if (
-              entry
-              && entry.memoryUsage
-              && typeof entry.memoryUsage === 'object'
-              && typeof entry.memoryUsage.label === 'string'
-              && entry.memoryUsage.label.trim() !== ''
-            ) {
-              return entry.memoryUsage.label.trim();
-            }
-            const formatted = formatPercentageValue(value);
-            if (formatted && formatted.trim() !== '') {
-              return formatted;
-            }
-            if (typeof value === 'number' && Number.isFinite(value)) {
-              return `${value.toFixed(1)} %`;
-            }
-            return 'Brak danych';
-          },
-        },
-        {
-          id: 'diskUsage',
-          label: 'Użycie dysku',
-          axisLabel: 'Użycie dysku (%)',
-          description: 'Wizualizacja historii zajętości dysku.',
-          formatValue: formatPercentageValue,
-          getValue: (entry) => {
-            if (!entry || typeof entry !== 'object' || !entry.diskUsage || typeof entry.diskUsage !== 'object') {
-              return null;
-            }
-            const value = entry.diskUsage.percentage;
-            return typeof value === 'number' && Number.isFinite(value) ? value : null;
-          },
-          getTooltipLabel: (entry, value) => {
-            if (
-              entry
-              && entry.diskUsage
-              && typeof entry.diskUsage === 'object'
-              && typeof entry.diskUsage.label === 'string'
-              && entry.diskUsage.label.trim() !== ''
-            ) {
-              return entry.diskUsage.label.trim();
-            }
-            const formatted = formatPercentageValue(value);
-            if (formatted && formatted.trim() !== '') {
-              return formatted;
-            }
-            if (typeof value === 'number' && Number.isFinite(value)) {
-              return `${value.toFixed(1)} %`;
-            }
-            return 'Brak danych';
-          },
-        },
-        {
-          id: 'systemLoad',
-          label: 'Obciążenie systemu',
-          axisLabel: 'Obciążenie (1 min)',
-          description: 'Wizualizacja historii obciążenia systemu (średnia z ostatniej minuty).',
-          formatValue: formatLoadAverageValue,
-          getValue: (entry) => {
-            if (!entry || typeof entry !== 'object' || !entry.systemLoad || typeof entry.systemLoad !== 'object') {
-              return null;
-            }
-            const value = entry.systemLoad.one;
-            return typeof value === 'number' && Number.isFinite(value) ? value : null;
-          },
-          getTooltipLabel: (entry, value) => {
-            if (
-              entry
-              && entry.systemLoad
-              && typeof entry.systemLoad === 'object'
-              && typeof entry.systemLoad.label === 'string'
-              && entry.systemLoad.label.trim() !== ''
-            ) {
-              return entry.systemLoad.label.trim();
-            }
-            const formatted = formatLoadAverageValue(value);
-            if (formatted && formatted.trim() !== '') {
-              return `1 min: ${formatted}`;
-            }
-            if (typeof value === 'number' && Number.isFinite(value)) {
-              return `1 min: ${value.toFixed(2)}`;
-            }
-            return 'Brak danych';
-          },
-        },
-      ];
-
-      const HISTORY_METRIC_MAP = HISTORY_METRICS.reduce((map, metric) => {
-        map[metric.id] = metric;
-        return map;
-      }, {});
-
-      const getHistoryMetricDefinition = (metricId) => HISTORY_METRIC_MAP[metricId] || HISTORY_METRICS[0];
-
-      const formatMetricValue = (metric, value) => {
-        if (!metric) {
-          return typeof value === 'number' && Number.isFinite(value) ? String(value) : '';
-        }
-        const formatted = metric.formatValue(value);
-        if (typeof formatted === 'string' && formatted.trim() !== '') {
-          return formatted;
-        }
-        if (typeof value === 'number' && Number.isFinite(value)) {
-          return String(value);
-        }
-        return '';
-      };
-
-      const setHistoryMetricState = (metricDefinition, options = {}) => {
-        const { persist = false } = options;
-        if (!metricDefinition || typeof metricDefinition !== 'object') {
-          return { changed: false, label: 'Metryka' };
-        }
-
-        const label = typeof metricDefinition.label === 'string' && metricDefinition.label.trim() !== ''
-          ? metricDefinition.label.trim()
-          : 'Metryka';
-        const previousMetric = historyState.metric;
-        const normalizedId = typeof metricDefinition.id === 'string' && metricDefinition.id.trim() !== ''
-          ? metricDefinition.id.trim()
-          : HISTORY_DEFAULT_METRIC;
-
-        historyState.metric = normalizedId;
-
-        const metricChanged = previousMetric !== normalizedId;
-        if (metricChanged) {
-          historyState.chartSignature = null;
-        }
-
-        if (persist) {
-          writeHistoryMetricPreference(normalizedId);
-        }
-
-        if (elements.historyChart) {
-          elements.historyChart.setAttribute('aria-label', `Historia: ${label}`);
-        }
-
-        if (elements.historyTitle) {
-          elements.historyTitle.textContent = `Historia: ${label}`;
-        }
-
-        const metricButtons = Array.from(elements.historyMetricButtons || []);
-        metricButtons.forEach((button) => {
-          if (!button || typeof button !== 'object') {
-            return;
-          }
-          button.setAttribute('type', 'button');
-          const target = button.getAttribute('data-metric');
-          const isActive = target === normalizedId;
-          button.classList.toggle('is-active', isActive);
-          button.setAttribute('aria-pressed', isActive ? 'true' : 'false');
-        });
-
-        return { changed: metricChanged, label };
-      };
-
-      const changeHistoryMetric = (metricId, persist = true) => {
-        const metricDefinition = getHistoryMetricDefinition(metricId);
-        setHistoryMetricState(metricDefinition, { persist });
-        updateHistoryChart();
-      };
-
-      const initializeHistoryMetricSwitch = () => {
-        const storedPreference = readHistoryMetricPreference();
-        const metricDefinition = getHistoryMetricDefinition(storedPreference || historyState.metric);
-        setHistoryMetricState(metricDefinition);
-
-        const metricButtons = Array.from(elements.historyMetricButtons || []);
-        if (metricButtons.length === 0) {
-          return;
-        }
-
-        metricButtons.forEach((button) => {
-          if (!button || typeof button !== 'object') {
-            return;
-          }
-          button.setAttribute('type', 'button');
-          button.addEventListener('click', (event) => {
-            event.preventDefault();
-            const target = button.getAttribute('data-metric');
-            if (typeof target !== 'string' || target.trim() === '') {
-              return;
-            }
-            changeHistoryMetric(target.trim(), true);
-          });
-        });
-      };
-
-      const renderHistoryChart = (entries, metricId) => {
-        const svg = elements.historyChart;
-        if (!svg) {
-          return false;
-        }
-        if (typeof SVGElement !== 'undefined' && !(svg instanceof SVGElement)) {
-          return false;
-        }
-
-        const metric = getHistoryMetricDefinition(metricId);
-        if (!metric) {
-          return false;
-        }
-
-        svg.setAttribute('viewBox', `0 0 ${CHART_VIEWBOX_WIDTH} ${CHART_VIEWBOX_HEIGHT}`);
-        svg.setAttribute('preserveAspectRatio', 'xMidYMid meet');
-        svg.innerHTML = '';
-
-        if (!Array.isArray(entries) || entries.length === 0) {
-          return false;
-        }
-
-        const values = entries.map((entry) => metric.getValue(entry));
-        if (values.some((value) => typeof value !== 'number' || !Number.isFinite(value))) {
-          return false;
-        }
-
-        const minValue = Math.min(...values);
-        const maxValue = Math.max(...values);
-
-        if (!Number.isFinite(minValue) || !Number.isFinite(maxValue)) {
-          return false;
-        }
-
-        let chartMin = minValue;
-        let chartMax = maxValue;
-
-        if (chartMax === chartMin) {
-          const offset = chartMax === 0 ? 1 : Math.abs(chartMax) * 0.1;
-          chartMin -= offset;
-          chartMax += offset;
-        } else {
-          const paddingValue = (chartMax - chartMin) * 0.05;
-          chartMin -= paddingValue;
-          chartMax += paddingValue;
-        }
-
-        const range = chartMax - chartMin;
-        if (range <= 0) {
-          return false;
-        }
-
-        const width = CHART_VIEWBOX_WIDTH;
-        const height = CHART_VIEWBOX_HEIGHT;
-        const padding = CHART_PADDING;
-        const innerWidth = width - padding.left - padding.right;
-        const innerHeight = height - padding.top - padding.bottom;
-
-        if (innerWidth <= 0 || innerHeight <= 0) {
-          return false;
-        }
-
-        const baseY = padding.top + innerHeight;
-        const baseXStart = padding.left;
-        const baseXEnd = width - padding.right;
-
-        const desc = createSvgElement('desc');
-        desc.textContent = metric.description || 'Wizualizacja historii metryki.';
-        svg.appendChild(desc);
-
-        const gridGroup = createSvgElement('g', { class: 'history-grid' });
-        for (let i = 0; i <= CHART_GRID_LINES; i += 1) {
-          const ratio = i / CHART_GRID_LINES;
-          const y = padding.top + (innerHeight * ratio);
-          const line = createSvgElement('line', {
-            x1: baseXStart,
-            y1: y,
-            x2: baseXEnd,
-            y2: y,
-            class: 'history-grid-line',
-          });
-          gridGroup.appendChild(line);
-
-          const value = chartMax - (range * ratio);
-          const label = createSvgElement('text', {
-            x: baseXStart - 8,
-            y,
-            class: 'history-axis-label history-axis-label--y',
-            'text-anchor': 'end',
-            'dominant-baseline': 'middle',
-          });
-          label.textContent = formatMetricValue(metric, value);
-          gridGroup.appendChild(label);
-        }
-        svg.appendChild(gridGroup);
-
-        const axisGroup = createSvgElement('g', { class: 'history-axis' });
-        const axisY = createSvgElement('line', {
-          x1: baseXStart,
-          y1: padding.top,
-          x2: baseXStart,
-          y2: baseY,
-          class: 'history-axis-line',
-        });
-        axisGroup.appendChild(axisY);
-
-        const axisX = createSvgElement('line', {
-          x1: baseXStart,
-          y1: baseY,
-          x2: baseXEnd,
-          y2: baseY,
-          class: 'history-axis-line',
-        });
-        axisGroup.appendChild(axisX);
-
-        const axisTitle = createSvgElement('text', {
-          x: baseXStart,
-          y: padding.top - 10,
-          class: 'history-axis-label history-axis-label--title',
-          'text-anchor': 'start',
-        });
-        axisTitle.textContent = metric.axisLabel || metric.label || 'Metryka';
-        axisGroup.appendChild(axisTitle);
-
-        const summary = createSvgElement('text', {
-          x: baseXEnd,
-          y: padding.top - 10,
-          class: 'history-axis-label history-axis-label--summary',
-          'text-anchor': 'end',
-        });
-        summary.textContent = `Min: ${formatMetricValue(metric, minValue)} · Max: ${formatMetricValue(metric, maxValue)}`;
-        axisGroup.appendChild(summary);
-
-        svg.appendChild(axisGroup);
-
-        const points = entries.map((entry, index) => {
-          const ratio = entries.length > 1 ? index / (entries.length - 1) : 0.5;
-          const x = padding.left + (innerWidth * ratio);
-          const value = values[index];
-          let normalized = (value - chartMin) / range;
-          if (!Number.isFinite(normalized)) {
-            normalized = 0;
-          }
-          normalized = Math.max(0, Math.min(1, normalized));
-          const y = padding.top + innerHeight - (normalized * innerHeight);
-          return { x, y, entry, value };
-        });
-
-        let areaPath = `M ${points[0].x.toFixed(2)} ${baseY.toFixed(2)}`;
-        let linePath = '';
-        points.forEach((point, index) => {
-          linePath += `${index === 0 ? 'M' : 'L'}${point.x.toFixed(2)},${point.y.toFixed(2)} `;
-          areaPath += ` L ${point.x.toFixed(2)} ${point.y.toFixed(2)}`;
-        });
-        areaPath += ` L ${points[points.length - 1].x.toFixed(2)} ${baseY.toFixed(2)} Z`;
-
-        const area = createSvgElement('path', {
-          d: areaPath.trim(),
-          class: 'history-area',
-        });
-        svg.appendChild(area);
-
-        const line = createSvgElement('path', {
-          d: linePath.trim(),
-          class: 'history-line',
-        });
-        svg.appendChild(line);
-
-        const dotsGroup = createSvgElement('g', { class: 'history-points' });
-        const radius = entries.length > 160 ? 1.2 : entries.length > 80 ? 1.6 : 2.4;
-        points.forEach((point) => {
-          const circle = createSvgElement('circle', {
-            cx: point.x.toFixed(2),
-            cy: point.y.toFixed(2),
-            r: radius,
-            class: 'history-dot',
-          });
-          const title = createSvgElement('title');
-          const tooltipLabel = metric.getTooltipLabel(point.entry, point.value);
-          const tooltip = typeof tooltipLabel === 'string' && tooltipLabel.trim() !== ''
-            ? tooltipLabel.trim()
-            : (formatMetricValue(metric, point.value) || 'Brak danych');
-          title.textContent = `${tooltip} (${formatHistoryEntryTime(point.entry)})`;
-          circle.appendChild(title);
-          dotsGroup.appendChild(circle);
-        });
-        svg.appendChild(dotsGroup);
-
-        const xLabelIndices = Array.from(new Set([
-          0,
-          entries.length > 1 ? Math.floor((entries.length - 1) / 2) : 0,
-          entries.length - 1,
-        ])).filter((index) => index >= 0 && index < entries.length).sort((a, b) => a - b);
-
-        const labelsGroup = createSvgElement('g', { class: 'history-x-labels' });
-        xLabelIndices.forEach((index) => {
-          const entry = entries[index];
-          const text = createSvgElement('text', {
-            x: points[index].x.toFixed(2),
-            y: (baseY + 18).toFixed(2),
-            class: 'history-axis-label history-axis-label--x',
-            'text-anchor': 'middle',
-            'dominant-baseline': 'hanging',
-          });
-          text.textContent = formatHistoryEntryTime(entry);
-          labelsGroup.appendChild(text);
-        });
-        svg.appendChild(labelsGroup);
-
-        return true;
-      };
-
-      const parseTemperatureValue = (value) => {
-        if (typeof value === 'number' && Number.isFinite(value)) {
-          return value;
-        }
-        if (typeof value !== 'string') {
-          return null;
-        }
-        const match = value.match(/(-?\d+(?:[\.,]\d+)?)/);
-        if (!match) {
-          return null;
-        }
-        const normalized = match[1].replace(',', '.');
-        const parsed = Number.parseFloat(normalized);
-        return Number.isFinite(parsed) ? parsed : null;
-      };
-
-      const parsePercentageValue = (value) => {
-        if (typeof value === 'number' && Number.isFinite(value)) {
-          return value;
-        }
-        if (typeof value !== 'string') {
-          return null;
-        }
-        const match = value.match(/(-?\d+(?:[\.,]\d+)?)\s*%/);
-        if (!match) {
-          return null;
-        }
-        const normalized = match[1].replace(',', '.');
-        const parsed = Number.parseFloat(normalized);
-        return Number.isFinite(parsed) ? parsed : null;
-      };
-
-      const parseSystemLoadValues = (value) => {
-        const result = { one: null, five: null, fifteen: null };
-        if (typeof value !== 'string') {
-          return result;
-        }
-        const trimmed = value.trim();
-        if (trimmed === '') {
-          return result;
-        }
-
-        const matchValue = (pattern) => {
-          const match = trimmed.match(pattern);
-          if (!match) {
-            return null;
-          }
-          const normalized = match[1].replace(',', '.');
-          const parsed = Number.parseFloat(normalized);
-          return Number.isFinite(parsed) ? parsed : null;
-        };
-
-        result.one = matchValue(/1\s*min\s*:\s*([\d\.,]+)/i);
-        result.five = matchValue(/5\s*min\s*:\s*([\d\.,]+)/i);
-        result.fifteen = matchValue(/15\s*min\s*:\s*([\d\.,]+)/i);
-
-        if (result.one === null || result.five === null || result.fifteen === null) {
-          const matches = trimmed.match(/(-?\d+(?:[\.,]\d+)?)/g);
-          if (matches && matches.length >= 3) {
-            const numbers = matches.slice(0, 3).map((item) => {
-              const normalizedNumber = item.replace(',', '.');
-              const parsed = Number.parseFloat(normalizedNumber);
-              return Number.isFinite(parsed) ? parsed : null;
-            });
-            const [first, second, third] = numbers;
-            if (result.one === null && typeof first === 'number' && Number.isFinite(first)) {
-              result.one = first;
-            }
-            if (result.five === null && typeof second === 'number' && Number.isFinite(second)) {
-              result.five = second;
-            }
-            if (result.fifteen === null && typeof third === 'number' && Number.isFinite(third)) {
-              result.fifteen = third;
-            }
-          }
-        }
-
-        return result;
-      };
-
-      const normalizePercentageMetric = (entry, key) => {
-        const result = { percentage: null, label: null };
-
-        if (!entry || typeof entry !== 'object') {
-          return result;
-        }
-
-        const rawMetric = entry[key];
-
-        if (rawMetric && typeof rawMetric === 'object' && !Array.isArray(rawMetric)) {
-          if (typeof rawMetric.label === 'string' && rawMetric.label.trim() !== '') {
-            result.label = rawMetric.label.trim();
-          }
-          if (typeof rawMetric.percentage === 'number' && Number.isFinite(rawMetric.percentage)) {
-            result.percentage = rawMetric.percentage;
-          }
-        } else if (typeof rawMetric === 'string' && rawMetric.trim() !== '') {
-          result.label = rawMetric.trim();
-        } else if (typeof rawMetric === 'number' && Number.isFinite(rawMetric)) {
-          result.percentage = rawMetric;
-        }
-
-        const labelKey = `${key}Label`;
-        if (result.label === null && typeof entry[labelKey] === 'string' && entry[labelKey].trim() !== '') {
-          result.label = entry[labelKey].trim();
-        }
-
-        const percentageKey = `${key}Percentage`;
-        const rawPercentage = entry[percentageKey];
-        if (result.percentage === null && typeof rawPercentage === 'number' && Number.isFinite(rawPercentage)) {
-          result.percentage = rawPercentage;
-        } else if (result.percentage === null && typeof rawPercentage === 'string' && rawPercentage.trim() !== '') {
-          const parsed = Number.parseFloat(rawPercentage.replace(',', '.'));
-          if (Number.isFinite(parsed)) {
-            result.percentage = parsed;
-          }
-        }
-
-        if (result.percentage === null && result.label !== null) {
-          const parsedFromLabel = parsePercentageValue(result.label);
-          if (Number.isFinite(parsedFromLabel)) {
-            result.percentage = parsedFromLabel;
-          }
-        }
-
-        if (result.label === null && result.percentage !== null) {
-          const formatted = formatPercentageValue(result.percentage);
-          result.label = formatted && formatted.trim() !== '' ? formatted : null;
-        }
-
-        return result;
-      };
-
-      const normalizeSystemLoadMetric = (entry) => {
-        const result = {
-          one: null,
-          five: null,
-          fifteen: null,
-          label: null,
-        };
-
-        if (!entry || typeof entry !== 'object') {
-          return result;
-        }
-
-        const rawMetric = entry.systemLoad;
-
-        if (rawMetric && typeof rawMetric === 'object' && !Array.isArray(rawMetric)) {
-          if (typeof rawMetric.label === 'string' && rawMetric.label.trim() !== '') {
-            result.label = rawMetric.label.trim();
-          }
-          if (typeof rawMetric.one === 'number' && Number.isFinite(rawMetric.one)) {
-            result.one = rawMetric.one;
-          }
-          if (typeof rawMetric.five === 'number' && Number.isFinite(rawMetric.five)) {
-            result.five = rawMetric.five;
-          }
-          if (typeof rawMetric.fifteen === 'number' && Number.isFinite(rawMetric.fifteen)) {
-            result.fifteen = rawMetric.fifteen;
-          }
-        } else if (typeof rawMetric === 'string' && rawMetric.trim() !== '') {
-          result.label = rawMetric.trim();
-        }
-
-        if (result.label === null && typeof entry.systemLoadLabel === 'string' && entry.systemLoadLabel.trim() !== '') {
-          result.label = entry.systemLoadLabel.trim();
-        }
-
-        if (
-          (result.one === null || result.five === null || result.fifteen === null)
-          && entry.systemLoadValues
-          && typeof entry.systemLoadValues === 'object'
-          && entry.systemLoadValues !== null
-        ) {
-          const values = entry.systemLoadValues;
-          if (result.one === null && typeof values.one === 'number' && Number.isFinite(values.one)) {
-            result.one = values.one;
-          }
-          if (result.five === null && typeof values.five === 'number' && Number.isFinite(values.five)) {
-            result.five = values.five;
-          }
-          if (result.fifteen === null && typeof values.fifteen === 'number' && Number.isFinite(values.fifteen)) {
-            result.fifteen = values.fifteen;
-          }
-        }
-
-        if (result.label !== null) {
-          const parsed = parseSystemLoadValues(result.label);
-          if (result.one === null && parsed.one !== null) {
-            result.one = parsed.one;
-          }
-          if (result.five === null && parsed.five !== null) {
-            result.five = parsed.five;
-          }
-          if (result.fifteen === null && parsed.fifteen !== null) {
-            result.fifteen = parsed.fifteen;
-          }
-        }
-
-        if (result.label === null) {
-          const parts = [];
-          if (typeof result.one === 'number' && Number.isFinite(result.one)) {
-            parts.push(`1 min: ${formatLoadAverageValue(result.one)}`);
-          }
-          if (typeof result.five === 'number' && Number.isFinite(result.five)) {
-            parts.push(`5 min: ${formatLoadAverageValue(result.five)}`);
-          }
-          if (typeof result.fifteen === 'number' && Number.isFinite(result.fifteen)) {
-            parts.push(`15 min: ${formatLoadAverageValue(result.fifteen)}`);
-          }
-          if (parts.length > 0) {
-            result.label = parts.join(' • ');
-          }
-        }
-
-        return result;
-      };
-
-      const createHistoryEntry = (raw) => {
-        if (!raw || typeof raw !== 'object') {
-          return null;
-        }
-
-        const generatedAt = typeof raw.generatedAt === 'string' ? raw.generatedAt : null;
-        const timeLabel = typeof raw.time === 'string'
-          ? raw.time
-          : (generatedAt ? new Date(generatedAt).toLocaleTimeString() : null);
-
-        let label = null;
-        let value = null;
-
-        if (raw.cpuTemperature && typeof raw.cpuTemperature === 'object' && !Array.isArray(raw.cpuTemperature)) {
-          if (typeof raw.cpuTemperature.label === 'string' && raw.cpuTemperature.label.trim() !== '') {
-            label = raw.cpuTemperature.label.trim();
-          }
-          if (typeof raw.cpuTemperature.value === 'number' && Number.isFinite(raw.cpuTemperature.value)) {
-            value = raw.cpuTemperature.value;
-          }
-        }
-
-        if (value === null && typeof raw.cpuTemperatureValue === 'number' && Number.isFinite(raw.cpuTemperatureValue)) {
-          value = raw.cpuTemperatureValue;
-        }
-
-        if (label === null && typeof raw.cpuTemperatureLabel === 'string' && raw.cpuTemperatureLabel.trim() !== '') {
-          label = raw.cpuTemperatureLabel.trim();
-        }
-
-        if (value === null && typeof raw.cpuTemperature === 'string') {
-          const parsed = parseTemperatureValue(raw.cpuTemperature);
-          if (Number.isFinite(parsed)) {
-            value = parsed;
-          }
-          if (!label && raw.cpuTemperature.trim() !== '') {
-            label = raw.cpuTemperature.trim();
-          }
-        }
-
-        if (value === null && label !== null) {
-          const parsed = parseTemperatureValue(label);
-          if (Number.isFinite(parsed)) {
-            value = parsed;
-          }
-        }
-
-        if (value !== null && (!label || label.trim() === '')) {
-          label = `${value.toFixed(1)} °C`;
-        }
-
-        const memoryUsage = normalizePercentageMetric(raw, 'memoryUsage');
-        const diskUsage = normalizePercentageMetric(raw, 'diskUsage');
-        const systemLoad = normalizeSystemLoadMetric(raw);
-
-        const hasNumericValue = [
-          value,
-          memoryUsage.percentage,
-          diskUsage.percentage,
-          systemLoad.one,
-          systemLoad.five,
-          systemLoad.fifteen,
-        ].some((metricValue) => typeof metricValue === 'number' && Number.isFinite(metricValue));
-
-        if (!hasNumericValue) {
-          return null;
-        }
-
-        return {
-          generatedAt,
-          time: timeLabel,
-          cpuTemperatureValue: value,
-          cpuTemperatureLabel: label && label.trim() !== '' ? label : null,
-          memoryUsage,
-          diskUsage,
-          systemLoad,
-        };
-      };
-
-      const updateHistoryChart = () => {
-        if (!elements.historyContainer) {
-          return;
-        }
-
-        const hideChart = () => {
-          if (elements.historyChartWrapper) {
-            elements.historyChartWrapper.classList.remove('is-visible');
-          }
-        };
-
-        const metricDefinition = getHistoryMetricDefinition(historyState.metric);
-        const { label: metricLabel } = setHistoryMetricState(metricDefinition);
-        const historyLabel = typeof metricLabel === 'string' && metricLabel.trim() !== ''
-          ? metricLabel.trim()
-          : 'Metryka';
-
-        if (historyState.enabled === null) {
-          destroyHistoryChart();
-          hideChart();
-
-          setHistoryMessage(`Historia metryki „${historyLabel}” ładuje się...`);
-          return;
-        }
-
-        if (historyState.enabled === false) {
-          destroyHistoryChart();
-          hideChart();
-
-          setHistoryMessage(`Historia metryki „${historyLabel}” jest wyłączona. Skonfiguruj zmienne środowiskowe, aby ją włączyć.`);
-          return;
-        }
-
-        const validEntries = historyState.entries.filter((entry) => {
-          if (!entry || typeof entry !== 'object') {
-            return false;
-          }
-          const value = metricDefinition.getValue(entry);
-          return typeof value === 'number' && Number.isFinite(value);
-        });
-
-        if (validEntries.length === 0) {
-          destroyHistoryChart();
-          hideChart();
-
-          setHistoryMessage(`Brak danych dla metryki „${historyLabel}”.`);
-          return;
-        }
-
-        const signature = JSON.stringify({
-          metric: metricDefinition.id,
-          entries: validEntries.map((entry) => {
-            const value = metricDefinition.getValue(entry);
-            const tooltipLabel = metricDefinition.getTooltipLabel(entry, value);
-            const tooltip = typeof tooltipLabel === 'string'
-              ? tooltipLabel.trim().slice(0, 160)
-              : '';
-            return [
-              entry.generatedAt ?? entry.time ?? '',
-              value,
-              tooltip,
-            ];
-          }),
-        });
-
-        if (historyState.chartSignature !== signature) {
-          const success = renderHistoryChart(validEntries, metricDefinition.id);
-          if (!success) {
-            destroyHistoryChart();
-            hideChart();
-            setHistoryMessage('Nie udało się narysować wykresu historii.');
-            return;
-          }
-          historyState.chartSignature = signature;
-
-        }
-
-        setHistoryMessage('', true);
-
-        if (elements.historyChartWrapper) {
-          elements.historyChartWrapper.classList.add('is-visible');
-        }
-      };
-
-      const pushSnapshotToHistory = (snapshot) => {
-        if (historyState.enabled !== true) {
-          return;
-        }
-
-        const entry = createHistoryEntry(snapshot);
-        if (!entry) {
-          return;
-        }
-
-        const lastEntry = historyState.entries.length > 0
-          ? historyState.entries[historyState.entries.length - 1]
-          : null;
-
-        if (lastEntry && entry.generatedAt && lastEntry.generatedAt === entry.generatedAt) {
-          historyState.entries[historyState.entries.length - 1] = entry;
-        } else {
-          historyState.entries.push(entry);
-        }
-
-        trimHistoryEntries();
-        updateHistoryChart();
-      };
-
-      const loadHistory = async () => {
-        if (!supportsFetch) {
-          historyState.enabled = false;
-          updateHistoryChart();
-          return;
-        }
-
-        if (!historyState.entries.length) {
-          setHistoryMessage('Ładowanie historii...');
-        }
-
-        try {
-          const response = await fetch(STATUS_HISTORY_ENDPOINT, {
-            headers: { Accept: 'application/json' },
-            cache: 'no-store',
-          });
-
-          if (!response.ok) {
-            throw new Error(`Błąd ${response.status}`);
-          }
-
-          const data = await response.json();
-
-          historyState.enabled = typeof data.enabled === 'boolean' ? data.enabled : true;
-          historyState.maxEntries = typeof data.maxEntries === 'number' && Number.isFinite(data.maxEntries)
-            ? data.maxEntries
-            : historyState.maxEntries;
-          historyState.limit = typeof data.limit === 'number' && Number.isFinite(data.limit)
-            ? data.limit
-            : historyState.limit;
-
-          if (Array.isArray(data.entries)) {
-            historyState.entries = data.entries
-              .map((entry) => createHistoryEntry(entry))
-              .filter((entry) => entry !== null);
-          } else {
-            historyState.entries = [];
-          }
-
-          trimHistoryEntries();
-          updateHistoryChart();
-        } catch (error) {
-          destroyHistoryChart();
-          if (elements.historyChartWrapper) {
-            elements.historyChartWrapper.classList.remove('is-visible');
-          }
-          const message = error instanceof Error ? error.message : String(error);
-          setHistoryMessage(`Nie udało się załadować historii: ${message}`);
-        }
-      };
-
-      const updateLabelSuccess = (generatedAt, serverTime, mode) => {
-        if (!elements.refreshLabel) {
-          return;
-        }
-
-        const timestamp = generatedAt
-          ? new Date(generatedAt).toLocaleTimeString()
-          : new Date().toLocaleTimeString();
-
-        const modeLabel = mode === 'stream' ? 'tryb na żywo' : (mode === 'poll' ? 'tryb zapasowy' : 'odświeżenie');
-        const serverLabel = serverTime ? ` (czas serwera: ${serverTime})` : '';
-        const intervalLabel = mode === 'stream'
-          && typeof STREAM_INTERVAL_SECONDS === 'number'
-          && Number.isFinite(STREAM_INTERVAL_SECONDS)
-          && STREAM_INTERVAL_SECONDS > 0
-            ? ` (interwał strumienia: ${STREAM_INTERVAL_SECONDS} s)`
-            : '';
-
-        elements.refreshLabel.textContent = `Ostatnia aktualizacja (${modeLabel}): ${timestamp}${serverLabel}${intervalLabel}`;
-      };
-
-      const renderServices = (services) => {
-        if (!elements.serviceList || !Array.isArray(services)) {
-          return;
-        }
-
-        const fragment = document.createDocumentFragment();
-
-        services.forEach((service) => {
-          if (!service || typeof service !== 'object') {
-            return;
-          }
-
-          const item = document.createElement('li');
-          const cssClass = typeof service.class === 'string' && service.class.trim() !== ''
-            ? service.class.trim()
-            : 'status-unknown';
-
-          item.className = `service-item ${cssClass}`;
-
-          if (service.details && typeof service.details === 'string' && service.details.trim() !== '') {
-            item.title = service.details;
-          }
-
-          const name = document.createElement('span');
-          name.className = 'service-name';
-          name.appendChild(document.createTextNode(fallback(service.label)));
-
-          const small = document.createElement('small');
-          if (typeof service.service === 'string' && service.service.trim() !== '') {
-            small.textContent = service.service;
-          } else {
-            small.textContent = '—';
-          }
-          name.appendChild(small);
-
-          const status = document.createElement('span');
-          status.className = 'service-status';
-          status.textContent = fallback(service.status);
-
-          item.appendChild(name);
-          item.appendChild(status);
-
-          fragment.appendChild(item);
-        });
-
-        elements.serviceList.innerHTML = '';
-        elements.serviceList.appendChild(fragment);
-      };
-
-      const applySnapshot = (data, mode = 'stream') => {
-        if (!data || typeof data !== 'object') {
-          return;
-        }
-
-        const serverTime = fallback(data.time ?? null);
-
-        if (elements.time) {
-          elements.time.textContent = serverTime;
-        }
-
-        if (elements.cpuTemperature) {
-          elements.cpuTemperature.textContent = fallback(data.cpuTemperature ?? null);
-        }
-
-        if (elements.memoryUsage) {
-          elements.memoryUsage.textContent = fallback(data.memoryUsage ?? null);
-        }
-
-        if (elements.diskUsage) {
-          elements.diskUsage.textContent = fallback(data.diskUsage ?? null);
-        }
-
-        if (elements.systemLoad) {
-          elements.systemLoad.textContent = fallback(data.systemLoad ?? null);
-        }
-
-        if (elements.uptime) {
-          elements.uptime.textContent = fallback(data.uptime ?? null);
-        }
-
-        renderServices(data.services);
-
-        updateLabelSuccess(data.generatedAt ?? null, serverTime, mode);
-
-        if (elements.refreshContainer) {
-          elements.refreshContainer.classList.remove('has-error');
-        }
-
-        pushSnapshotToHistory(data);
-      };
-
-      const loadStatus = async (manual = false) => {
-        if (!supportsFetch || isFetching) {
-          return;
-        }
-
-        isFetching = true;
-        setFetchingState(true, manual);
-
-        try {
-          const response = await fetch(STATUS_JSON_ENDPOINT, {
-            headers: { Accept: 'application/json' },
-            cache: 'no-store',
-          });
-
-          if (!response.ok) {
-            throw new Error(`Błąd ${response.status}`);
-          }
-
-          const data = await response.json();
-          const mode = manual
-            ? (supportsEventSource && eventSource ? 'stream' : 'manual')
-            : (supportsEventSource && eventSource ? 'stream' : 'poll');
-
-          applySnapshot(data, mode);
-        } catch (error) {
-          if (elements.refreshContainer) {
-            elements.refreshContainer.classList.add('has-error');
-          }
-
-          if (elements.refreshLabel) {
-            const timestamp = new Date().toLocaleTimeString();
-            const message = error instanceof Error ? error.message : String(error);
-            elements.refreshLabel.textContent = `Błąd odświeżania (${timestamp}): ${message}`;
-          }
-        } finally {
-          isFetching = false;
-          setFetchingState(false);
-        }
-      };
-
-      const stopPolling = () => {
-        if (refreshTimer) {
-          clearInterval(refreshTimer);
-          refreshTimer = null;
-        }
-      };
-
-      const startPolling = () => {
-        if (!supportsFetch || refreshTimer) {
-          return;
-        }
-
-        loadStatus();
-        refreshTimer = window.setInterval(() => loadStatus(), REFRESH_INTERVAL);
-      };
-
-      const cancelReconnect = () => {
-        if (reconnectTimer) {
-          clearTimeout(reconnectTimer);
-          reconnectTimer = null;
-        }
-      };
-
-      const stopStream = () => {
-        if (eventSource) {
-          eventSource.close();
-          eventSource = null;
-        }
-        cancelReconnect();
-      };
-
-      const scheduleReconnect = () => {
-        if (reconnectTimer) {
-          return;
-        }
-
-        reconnectTimer = window.setTimeout(() => {
-          reconnectTimer = null;
-          startStream();
-        }, STREAM_RECONNECT_DELAY);
-      };
-
-      const startStream = () => {
-        if (!supportsEventSource) {
-          if (supportsFetch) {
-            startPolling();
-          }
-          return;
-        }
-
-        stopStream();
-        stopPolling();
-
-        if (elements.refreshContainer) {
-          elements.refreshContainer.classList.remove('has-error');
-        }
-
-        setFetchingState(true, false, 'Łączenie ze strumieniem danych...');
-
-        eventSource = new EventSource(STATUS_STREAM_ENDPOINT);
-
-        const handleStatusEvent = (event) => {
-          try {
-            const data = JSON.parse(event.data);
-            applySnapshot(data, 'stream');
-            setFetchingState(false);
-            stopPolling();
-          } catch (parseError) {
-            if (elements.refreshContainer) {
-              elements.refreshContainer.classList.add('has-error');
-            }
-            if (elements.refreshLabel) {
-              const timestamp = new Date().toLocaleTimeString();
-              elements.refreshLabel.textContent = `Błąd danych strumienia (${timestamp})`;
-            }
-          }
-        };
-
-        eventSource.addEventListener('status', handleStatusEvent);
-        eventSource.addEventListener('message', handleStatusEvent);
-
-        eventSource.addEventListener('open', () => {
-          setFetchingState(false);
-        });
-
-        eventSource.addEventListener('error', () => {
-          setFetchingState(false);
-          if (elements.refreshContainer) {
-            elements.refreshContainer.classList.add('has-error');
-          }
-          if (elements.refreshLabel) {
-            const timestamp = new Date().toLocaleTimeString();
-            elements.refreshLabel.textContent = `Błąd strumienia (${timestamp}). Próba ponownego połączenia...`;
-          }
-          stopStream();
-          if (supportsFetch) {
-            startPolling();
-          }
-          scheduleReconnect();
-        });
-      };
-
-      renderShellyDevices();
-
-      if (shellyConfigError) {
-        stopShellyAutoRefresh();
-        setShellyError(SHELLY_CONFIG_ERROR_MESSAGE);
-        setShellyMessage(SHELLY_CONFIG_ERROR_HINT);
-        if (elements.shellyList) {
-          elements.shellyList.innerHTML = '';
-          elements.shellyList.classList.remove('is-loading');
-          elements.shellyList.setAttribute('aria-busy', 'false');
-        }
-        if (elements.shellyReload) {
-          elements.shellyReload.disabled = true;
-          elements.shellyReload.setAttribute('aria-disabled', 'true');
-        }
-      } else {
-        if (elements.shellyList) {
-          elements.shellyList.addEventListener('click', handleShellyActionClick);
-        }
-
-        if (elements.shellyReload) {
-          elements.shellyReload.addEventListener('click', () => {
-            loadShellyDevices(true);
-          });
-
-          if (!supportsFetch) {
-            elements.shellyReload.disabled = true;
-            elements.shellyReload.setAttribute('aria-disabled', 'true');
-          } else {
-            elements.shellyReload.removeAttribute('aria-disabled');
-          }
-        }
-      }
-
-      initializeTabs();
-
-      if (elements.themeToggle) {
-        elements.themeToggle.addEventListener('click', () => {
-          const isDark = !document.body.classList.contains('theme-dark');
-          applyThemePreference(isDark, true);
-        });
-      }
-
-      initializeTheme();
-
-      if (elements.statusNote
-        && typeof STREAM_INTERVAL_SECONDS === 'number'
-        && Number.isFinite(STREAM_INTERVAL_SECONDS)
-        && STREAM_INTERVAL_SECONDS > 0
-      ) {
-        elements.statusNote.textContent = `Dane odświeżają się automatycznie (interwał: ${STREAM_INTERVAL_SECONDS} s). W przypadku problemów spróbuj kliknąć przycisk poniżej.`;
-      }
-
-      if (elements.refreshButton) {
-        elements.refreshButton.addEventListener('click', () => {
-          if (supportsFetch) {
-            loadStatus(true);
-            loadHistory();
-          } else if (supportsEventSource) {
-            stopStream();
-            startStream();
-          }
-        });
-      }
-
-      initializeHistoryMetricSwitch();
-      updateHistoryChart();
-
-      if (supportsFetch) {
-        loadHistory();
-      } else {
-        historyState.enabled = false;
-        destroyHistoryChart();
-        if (elements.historyChartWrapper) {
-          elements.historyChartWrapper.classList.remove('is-visible');
-        }
-        setHistoryMessage('Historia wymaga przeglądarki obsługującej funkcję fetch.');
-      }
-
-      if (supportsEventSource) {
-        startStream();
-      } else if (supportsFetch) {
-        startPolling();
-      }
-
-      document.addEventListener('visibilitychange', () => {
-        if (document.visibilityState === 'visible') {
-          if (supportsEventSource) {
-            if (!eventSource) {
-              startStream();
-            }
-          } else if (supportsFetch && !refreshTimer) {
-            startPolling();
-          }
-          if (currentTab === SHELLY_TAB_ID) {
-            ensureShellyDataLoaded(false);
-            startShellyAutoRefresh();
-          }
-        } else {
-          if (supportsEventSource) {
-            stopStream();
-          }
-          stopPolling();
-          stopShellyAutoRefresh();
-        }
-      });
-    })();
-  </script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
