+++ public/js/support.js
@@ -1,0 +1,24 @@
+export const supportsFetch = typeof window.fetch === 'function';
+export const supportsEventSource = typeof window.EventSource === 'function';
+
+export function parsePositiveInteger(value) {
+  if (typeof value === 'number' && Number.isFinite(value) && value > 0) {
+    return Math.floor(value);
+  }
+
+  if (typeof value !== 'string') {
+    return null;
+  }
+
+  const trimmed = value.trim();
+  if (trimmed === '') {
+    return null;
+  }
+
+  const parsed = Number.parseInt(trimmed, 10);
+  if (!Number.isFinite(parsed) || parsed <= 0) {
+    return null;
+  }
+
+  return parsed;
+}
