+++ public/styles.css
@@ -50,46 +50,46 @@ body {
 }
 
 body.theme-dark {
-  --background: #0f1724;
-  --text-primary: #e2e8f8;
-  --heading-primary: #8cb5ff;
-  --panel-background: #1b2535;
-  --panel-shadow: 0 12px 32px rgba(0, 0, 0, 0.45);
-  --panel-heading: #adc8ff;
-  --metric-background: #202c3f;
-  --metric-label: #a4b3d6;
-  --metric-value: #d9e6ff;
-  --service-item-background: #1f2a3d;
-  --service-name: #d9e6ff;
-  --service-name-muted: #9aa9c9;
-  --status-ok: #6fdc8c;
-  --status-off-border: #4c5874;
-  --status-off-text: #aab7d1;
-  --status-error: #ff7f76;
+  --background: #050505;
+  --text-primary: #f2f2f2;
+  --heading-primary: #fafafa;
+  --panel-background: #111111;
+  --panel-shadow: 0 12px 32px rgba(0, 0, 0, 0.7);
+  --panel-heading: #f5f5f5;
+  --metric-background: #161616;
+  --metric-label: #bdbdbd;
+  --metric-value: #ffffff;
+  --service-item-background: #151515;
+  --service-name: #f0f0f0;
+  --service-name-muted: #9c9c9c;
+  --status-ok: #4caf50;
+  --status-off-border: #2c2c2c;
+  --status-off-text: #b0b0b0;
+  --status-error: #ef5350;
   --status-warn-border: #f2c96b;
-  --status-warn-text: #f6e2a0;
-  --status-unknown: #91a1c2;
-  --status-note: #bbc6dd;
-  --refresh-text: #bbc6dd;
-  --refresh-error: #ff9b92;
-  --button-background: #3b5cab;
-  --button-background-hover: #4f6fca;
-  --button-background-disabled: #56658a;
-  --button-active-background: #274a8a;
-  --button-text: #f8fbff;
-  --footer-text: #a4b3d6;
-  --history-background: #202c3f;
-  --history-area: rgba(93, 131, 255, 0.25);
-  --history-line: #91b4ff;
-  --history-dot: #d2e0ff;
-  --history-dot-border: #1b2535;
-  --history-grid-line: #2f3c55;
-  --history-axis-line: #4d5d7a;
-  --history-axis-label: #b7c2db;
-  --history-axis-title: #e2e8f8;
-  --history-axis-x: #a4b3d6;
-  --panel-border: #2f3c55;
-  --outer-shadow: 0 0 16px rgba(0, 0, 0, 0.6);
+  --status-warn-text: #ffe082;
+  --status-unknown: #9e9e9e;
+  --status-note: #cccccc;
+  --refresh-text: #cccccc;
+  --refresh-error: #ff8a80;
+  --button-background: #1f1f1f;
+  --button-background-hover: #2a2a2a;
+  --button-background-disabled: #333333;
+  --button-active-background: #000000;
+  --button-text: #f5f5f5;
+  --footer-text: #bdbdbd;
+  --history-background: #121212;
+  --history-area: rgba(255, 255, 255, 0.1);
+  --history-line: #f5f5f5;
+  --history-dot: #ffffff;
+  --history-dot-border: #111111;
+  --history-grid-line: #2a2a2a;
+  --history-axis-line: #3a3a3a;
+  --history-axis-label: #c7c7c7;
+  --history-axis-title: #f5f5f5;
+  --history-axis-x: #bdbdbd;
+  --panel-border: #1f1f1f;
+  --outer-shadow: 0 0 20px rgba(0, 0, 0, 0.85);
 }
 
 h1 {
